Remove broken modification tracking
The issue we saw with ome/scripts#12 was:

 * PR 12 opened and was merged properly.
 * PR 12 was merged.
 * No other PRs were open.
 * Therefore, the sha1 of the module wasn't changed.

In other words, the modification tracking that was
in place looked solely for PR merges but didn't
take into account whether or not the submodule
pointer needed to be bumped. Rather than worry about
trying to properly track that, this commit removes
the modification check. The next step is to
guarantee that the submodule will be updated
to the appropriate base.