Add CleanSandbox command
The clean-sandbox target is used to remove
all branches not named "master" from the fork
of "snoopys-sandbox" for the current user.
This works similar to "git clean" in that either
`-n` or `-f` must be specified.

A minor refactoring of all Command.__call__
implementations took place to do basic setup
in the super() method.