Add UpdateSubmodules command
Due to the complexity of rmerge and co., a new command
with a similar signature to `scc merge` is preferred for
updating all submodules.

The idea is that for this to work, one must only change
to the appropriate branch, and run

    `scc update-submodules BASE`

All of the submodules will be fetched and checked out to
the appropriate base, then a single final commit will be
added.