Add UnrebasedPRs command
Command is used to compare two branches and look for
and merge commits which match the Github-style pattern.
Then, the git notes under ref "see_also/BRANCH" are
consulted to see if the PR has a rebase-pointer. If yes,
then it is ignored. Otherwise, the PR is printed to either
stdout or to a file. The two files from each invocation can
be used (via the --parse flag) to add notes so that they will
be filtered on subsequent invocations.