Add --shallow functionality
This function, added to all instances or subclasses of GitRepoCommand,
allows the user to stop scc from recursively resolving and operating on
the submodules of the current repository.  This is particularly
important when an integration branch is not present on all submodules or
where scc operations on one or more submodules is undesirable.