Add option to filter PRs by status in the merge command
This commit defines a new flag argument --check-commit-status for all commands
derived from the FilteredPullRequestCommand abstract class. The
find_candidates() method is updated to add an extra filtering step if this
flag is passed.
This commit provides the default framework to close #41. Extra granularity on
the status filters can be implemented when necessary.