Fix scc check-prs by using find_branching_point()
This commit addresses a tricky issue where a branch originated from an
ancestor commit is merged simultaneously in the source and target branches.
In that case, `git merge-base` will return this ancestor commit instead of
the true branching point. To fix this, CheckPRs now uses find_branching_point
already used in Rebase().