Add logging at ERROR after GitHubRepository.get_pull() exceptions
Without some form of logging at this or some other level in the stack
the sum total of information available following a failure to get access
to pull request metadata is:

    UnknownObjectException: 404 {u'documentation_url': u'https://developer.github.com/v3', u'message': u'Not Found'}

This is needless to say, insufficent.  Logging of failures will now
happen in the form:

    Failure to get pull request openmicroscopy/snoopys-sandbox#-1

With these changes it should be possible to debug missing repositories
or repositories for which the current user does not have access
permissions much more easily.