Always push the branch in the Merge command
The current design was implemented for repositories with lots of open requests
in mind so that the merge branch is always updated. However, for repositories
or development branches with few PRs opened, the requirement can mean the
merge branch never gets updated to the HEAD of the development branch if all
PRs are closed or excluded.

This commit removes this restriction by pushing unconditionally the merge
branch when the --push argument is passed.