Improve tracking branch filter parsing
While the filter branch functionality is working and checked by an integration
test, it fails to integrate a branch called `bf_5.3.0-SNAPSHOT`. This is
a deficiency in the regexp logic when identifying candidate branches.

To fix this issue, this PR extends the _parse_filters() logic to actively
parse tracking branches using regular expressions. Unit tests are also added
covering the range of tracking branch names.

Support is also added for branch tracking using the full Github URL similarly
to what is done for Pull Request inclusion.