Fix rate limiting calls following upstream API breakage
PyGithub 1.43 introduces a breaking change in the RateLimit API modifying both
the internal representation as well as the methods of the Rate objects - see https://github.com/PyGithub/PyGithub/commit/fd8a036. This causes scc to crash
when a recent version of PyGithub is installed.

Rather than capping PyGithub or forcing an upgrade, this PR adds
backwards-compatible support for the new rate limit API with the following
changes:
- the get_rate_limits() method is deprecated
- a new get_rate_limit() method forwards the calls to GitHub.get_rate_limit()
- all internal usages of rate limits are updated to use get_rate_limit. A
  try/exept block is added to handle older and newer version of PyGithub and
  return the core rate as well as the search and graphql rates for recent
  versions.